Changed: view.Writer renamed methods and updated docs.
